Rename test conversion
